Fixed spelling and grammar for clarity (#730)
* Fixed spelling and grammar for clarity

* Update README.adoc

* Three comma club

* gramma grammma grammar

* Update README.adoc

* Remove commas

Corrected the commas.

Consider the sentences:

- "Energy-saving light bulbs which are cheap make for a sensible investment"
- "Energy-saving light bulbs, which are cheap, make for a sensible investment"

They are different. The first one is intimating that energy-saving light bulbs come in many prices and it's only the cheap ones which are a good investment. The second implies that all energy-saving light bulbs are generally cheap and they, regardless of price, make for a good investment.

The commas added (which I just removed) switched the meaning, incorrectly, from one to the other.